Refactored code, added error handling and documentation

+++ backend/SquareOverFlowApi/SquareOverFlowCore/Models/Color.cs
@@ -2,8 +2,8 @@
 {
     public class Color
     {
-        public int Red { get; set; } = 0;
-        public int Green { get; set; } = 0;
-        public int Blue { get; set; } = 0;
+        public byte Red { get; set; } = 0;
+        public byte Green { get; set; } = 0;
+        public byte Blue { get; set; } = 0;
     }
 }
